update file stories
